feat: add episode group for solve issue #389. (#394)
* feat: add episode group for solve issue #389.

* docs: update changelog for pr #394.

* fix: checkstyle issue.

* fix: episode unit test case.

* build: gen new api-client@0.4.1 and publish to npm center repo in @runikaros/api-client

* fix: console build issue.

+++ server/src/main/java/run/ikaros/server/store/repository/EpisodeRepository.java
@@ -10,5 +10,5 @@ public interface EpisodeRepository extends R2dbcRepository<EpisodeEntity, Long> 
 
     Mono<Void> deleteAllBySubjectId(Long subjectId);
 
-    Mono<EpisodeEntity> findBySubjectIdAndSequence(Long subjectId, Double sequence);
+    Mono<EpisodeEntity> findBySubjectIdAndSequence(Long subjectId, Integer sequence);
 }
